Fix code review issues from search phase one
Deduplicate shared helpers across search UI states (startPlanPlayback,
startSearch, subscribeToBasicRender) and server states
(broadcastExecutionRender). Fix alphabetical ordering in ApiCommand,
Event, and UiApiCommand variants. Replace raw spdlog calls with SLOG
macros in PlanRepository. Move PlayerControlFrame to_json/from_json out
of header into .cpp. Remove unused spdlog include from
SmbPlanExecution.
