+++ apps/CMakeLists.txt
@@ -309,6 +309,7 @@ set(DIRTSIM_CORE_SOURCES
 
     # Gamepad input.
     src/core/input/GamepadManager.cpp
+    src/core/input/PlayerControlFrame.cpp
 )
 
 add_library(dirtsim-core STATIC ${DIRTSIM_CORE_SOURCES})
@@ -343,6 +344,7 @@ add_library(dirtsim-server-lib STATIC
     src/server/states/Idle.cpp
     src/server/states/PlanPlayback.cpp
     src/server/states/SearchActive.cpp
+    src/server/states/SearchBroadcastHelpers.cpp
     src/server/states/Shutdown.cpp
     src/server/states/SimPaused.cpp
     src/server/states/SimRunning.cpp
@@ -548,6 +550,7 @@ add_library(dirtsim-ui-lib STATIC
     src/ui/state-machine/states/Paused.cpp
     src/ui/state-machine/states/PlanPlayback.cpp
     src/ui/state-machine/states/SearchActive.cpp
+    src/ui/state-machine/states/SearchHelpers.cpp
     src/ui/state-machine/states/SearchIdle.cpp
     src/ui/state-machine/states/SearchPlanBrowser.cpp
     src/ui/state-machine/states/Shutdown.cpp
