+++ apps/src/core/input/PlayerControlFrame.h
@@ -25,20 +25,7 @@ struct PlayerControlFrame {
     using serialize = zpp::bits::members<3>;
 };
 
-inline void to_json(nlohmann::json& j, const PlayerControlFrame& value)
-{
-    j = nlohmann::json{
-        { "xAxis", value.xAxis },
-        { "yAxis", value.yAxis },
-        { "buttons", value.buttons },
-    };
-}
-
-inline void from_json(const nlohmann::json& j, PlayerControlFrame& value)
-{
-    j.at("xAxis").get_to(value.xAxis);
-    j.at("yAxis").get_to(value.yAxis);
-    j.at("buttons").get_to(value.buttons);
-}
+void to_json(nlohmann::json& j, const PlayerControlFrame& value);
+void from_json(const nlohmann::json& j, PlayerControlFrame& value);
 
 } // namespace DirtSim
