+++ apps/src/server/api/ApiCommand.h
@@ -124,11 +124,11 @@ using ApiCommand = std::variant<
     Api::RenderStreamConfigSet::Command,
     Api::Reset::Command,
     Api::ScenarioListGet::Command,
+    Api::ScenarioSwitch::Command,
     Api::SearchPauseSet::Command,
     Api::SearchProgressGet::Command,
     Api::SearchStart::Command,
     Api::SearchStop::Command,
-    Api::ScenarioSwitch::Command,
     Api::SeedAdd::Command,
     Api::SimRun::Command,
     Api::SimStop::Command,
