+++ apps/src/server/states/PlanPlayback.cpp
@@ -1,10 +1,9 @@
+#include "SearchBroadcastHelpers.h"
 #include "State.h"
 #include "core/LoggingChannels.h"
-#include "core/ScenarioConfig.h"
 #include "core/network/BinaryProtocol.h"
 #include "server/StateMachine.h"
 #include "server/api/PlanPlaybackStopped.h"
-#include <vector>
 
 namespace DirtSim {
 namespace Server {
@@ -25,21 +24,6 @@ void broadcastPlanPlaybackStopped(
     dsm.broadcastEventData(Api::PlanPlaybackStopped::name(), Network::serialize_payload(stopped));
 }
 
-void broadcastPlanPlaybackRender(
-    StateMachine& dsm, const SearchSupport::SmbPlanExecution& execution)
-{
-    static const std::vector<OrganismId> emptyOrganismGrid{};
-    const auto scenarioId = Scenario::EnumType::NesSuperMarioBros;
-    const auto scenarioConfig = makeDefaultConfig(scenarioId);
-    dsm.broadcastRenderMessage(
-        execution.getWorldData(),
-        emptyOrganismGrid,
-        scenarioId,
-        scenarioConfig,
-        std::nullopt,
-        execution.getScenarioVideoFrame());
-}
-
 } // namespace
 
 void PlanPlayback::onEnter(StateMachine& dsm)
@@ -48,7 +32,7 @@ void PlanPlayback::onEnter(StateMachine& dsm)
     dsm.updateCachedWorldData(execution.getWorldData());
     renderBroadcasted_ = false;
     if (execution.hasRenderableFrame()) {
-        broadcastPlanPlaybackRender(dsm, execution);
+        broadcastExecutionRender(dsm, execution);
         renderBroadcasted_ = true;
     }
 }
@@ -63,7 +47,7 @@ std::optional<Any> PlanPlayback::tick(StateMachine& dsm)
     const auto tickResult = execution.tick();
     if (execution.hasRenderableFrame() && (!renderBroadcasted_ || tickResult.frameAdvanced)) {
         dsm.updateCachedWorldData(execution.getWorldData());
-        broadcastPlanPlaybackRender(dsm, execution);
+        broadcastExecutionRender(dsm, execution);
         renderBroadcasted_ = true;
     }
