+++ apps/src/ui/state-machine/Event.h
@@ -399,10 +399,10 @@ using Event = std::variant<
     DirtSim::UiApi::PlanBrowserOpen::Cwc,
     DirtSim::UiApi::PlanDetailOpen::Cwc,
     DirtSim::UiApi::PlanDetailSelect::Cwc,
-    DirtSim::UiApi::PlantSeed::Cwc,
     DirtSim::UiApi::PlanPlaybackPauseSet::Cwc,
     DirtSim::UiApi::PlanPlaybackStart::Cwc,
     DirtSim::UiApi::PlanPlaybackStop::Cwc,
+    DirtSim::UiApi::PlantSeed::Cwc,
     DirtSim::UiApi::RenderModeSelect::Cwc,
     DirtSim::UiApi::ScreenGrab::Cwc,
     DirtSim::UiApi::SearchPauseSet::Cwc,
