+++ apps/src/ui/state-machine/api/UiApiCommand.h
@@ -77,10 +77,10 @@ using UiApiCommand = std::variant<
     UiApi::PlanBrowserOpen::Command,
     UiApi::PlanDetailOpen::Command,
     UiApi::PlanDetailSelect::Command,
-    UiApi::PlantSeed::Command,
     UiApi::PlanPlaybackPauseSet::Command,
     UiApi::PlanPlaybackStart::Command,
     UiApi::PlanPlaybackStop::Command,
+    UiApi::PlantSeed::Command,
     UiApi::RenderModeSelect::Command,
     UiApi::ScreenGrab::Command,
     UiApi::SearchPauseSet::Command,
